requirements.txt
asgiref==3.4.1
attrs==21.2.0
black==21.9b0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
et-xmlfile==1.1.0
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
iniconfig==1.1.1
motor==2.5.1
msgpack==1.0.2
mypy-extensions==0.4.3
numpy==1.21.2
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
pathspec==0.9.0
pendulum==2.1.2
platformdirs==2.3.0
pluggy==1.0.0
py==1.10.0
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.12.0
pynvim==0.4.3
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
pytzdata==2020.1
regex==2021.9.24
requests==2.26.0
six==1.16.0
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
